updated the dashboard for both users and osvaldy with using my email currently.
Here, I added the user dashboard. The user dashboard now has access to its components and page. I still need to design it and make it more functional and appealing. The admin dashboard has been established, and now it is time to implement design and function. Before i do, I will integrate square to obtain the ability to charge clients thru the page.